implemented "acknowledgement" feature
added ability to "acknowledge" waypoints by pinging an existing QC waypoint, causing a color-coded checkmark to appear next to that waypoint (corresponding to the peer color of the peer acknowledging it). this feature SHOULD be compatible with biglobby, though it is not tested.
note: gcw waypoints will be visibly acknowledged by any qc users, but not by gcw users, since gcw has no analogy or counterpart to this feature